+++ resources/views/v_warung.blade.php
@@ -51,10 +51,12 @@
                 <p class="warung-phone" id="warung-phone{{ $warung->id }}">{{ $warung->nohp }}</p>
                 <p id="warung-address{{ $warung->id }}">{{ $warung->alamat }}</p>
                 <p>Produk terbeli : {{ $warung->total }}</p>
-                <div class="action">
-                    <button type="button" class="edit" onclick="editWarung({{ $warung->id }})">EDIT</button>
-                    <button type="button" class="hapus" onclick="hapusWarung({{ $warung->id }})" >DELETE</button>
-                </div>
+                @if(!auth()->user()->is_admin)
+                    <div class="action">
+                        <button type="button" class="edit" onclick="editWarung({{ $warung->id }})">EDIT</button>
+                        <button type="button" class="hapus" onclick="hapusWarung({{ $warung->id }})" >DELETE</button>
+                    </div>
+                @endif
                 <a href="/pencatatan/{{ $warung->id }}/{{ idate('Y') }}/{{ idate('m') }}">PENCATATAN</a>
             </div>
         @endforeach
